pyproject.toml
[tool.poetry]
name = "ai-hedge-fund"
version = "0.1.0"
description = "An AI-powered hedge fund that uses multiple agents to make trading decisions"
authors = ["imasker <imasker@qq.com>"]
readme = "README.md"
packages = [
    { include = "src" }
]
[tool.poetry.dependencies]
python = "^3.9"
langchain = "0.3.0"
langchain-openai = "0.2.11"
langgraph = "0.2.56"
pandas = "^2.1.0"
numpy = "^1.24.0"
python-dotenv = "1.0.0"
matplotlib = "^3.9.2"
yfinance = "^0.2.51"
akshare = "^1.15.64"
requests = "^2.31.0"
beautifulsoup4 = "^4.12.3"
openai = "^1.12.0"
langchain-core = "^0.3.29"
google-generativeai = "^0.8.3"
backoff = "^2.2.1"
ollama = "0.4.6"

[tool.poetry.group.dev.dependencies]
pytest = "^7.4.0"
black = "^23.7.0"
isort = "^5.12.0"
flake8 = "^6.1.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"